# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MindMora auth API",
    "version": "0.1.0",
    "description": "Milestone 1B auth-only contract. Bounded live Google/session/logout and refresh/replay acceptance recorded in docs/phases/phase-01-foundation.md. Notes, Swagger UI, Postman generation and full contract drift tooling are not implemented. 1D adds bounded empty-body checks, correlation IDs, Redis admission and safe Pino logging; hosted readiness remains unverified."
  },
  "paths": {
...
    "/api/auth/callback": {
      "get": {
        "operationId": "authCallback",
        "summary": "Backend-owned auth callback",
        "description": "No caller-controlled redirect. Auth responses are no-store. 1D bounds empty bodies and applies Redis admission and Pino metadata. Hosted readiness remains unverified.",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{43}$"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect to the fixed APP_ORIGIN homepage; writes protected app cookie and removes pending cookie.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "const": "no-referrer"
                }
              },
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "Set-Cookie": {
                "description": "Host-only HttpOnly SameSite=Lax; Secure/__Host over HTTPS; no credential examples.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            }
          },
          "400": {
            "description": "Safe auth error; raw provider errors and credentials are excluded.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "const": "no-referrer"
                }
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "401": {
            "description": "Safe auth error; raw provider errors and credentials are excluded.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "const": "no-referrer"
                }
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "403": {
            "description": "Safe auth error; raw provider errors and credentials are excluded.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "const": "no-referrer"
                }
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "405": {
            "description": "Safe auth error; raw provider errors and credentials are excluded.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "const": "no-referrer"
                }
              },
              "Allow": {
                "schema": {
                  "type": "string",
                  "const": "GET"
                }
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "503": {
            "description": "Provider/config or Redis admission unavailable. Admission-rejected logout retains cookies and is unsuccessful.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string",
                  "const": "no-referrer"
                }
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "X-Content-Type-Options": {
                "$ref": "#/components/headers/NoSniff"
              },
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/BoundedRequest"
          },
          "429": {
            "$ref": "#/components/responses/AuthThrottled"
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedFailure"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AppSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-mindmora-session",
        "description": "Protected HTTPS app-session cookie; never copy credentials into docs."
      },
      "LocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mindmora-session",
        "description": "Loopback HTTP development/preview exemption only."
      }
    },
    "schemas": {
...
      "AuthError": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "correlationId"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "auth_unavailable",
                  "unauthenticated",
                  "origin_rejected",
                  "invalid_auth_request",
                  "invalid_auth_callback",
                  "method_not_allowed",
                  "invalid_request",
                  "payload_too_large",
                  "rate_limited",
                  "admission_unavailable",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              },
              "correlationId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      }
    },
    "headers": {
      "CorrelationId": {
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Server-generated request correlation ID."
      },
      "NoSniff": {
        "schema": {
          "type": "string",
          "const": "nosniff"
        }
      },
      "RetryAfter": {
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]*$"
        },
        "description": "Seconds until retry; present on admission rejections."
      }
    },
    "responses": {
      "BoundedRequest": {
        "description": "Request exceeds header, URL or body limits.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "private, no-store"
            }
          },
          "Referrer-Policy": {
            "schema": {
              "type": "string",
              "const": "no-referrer"
            }
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/CorrelationId"
          },
          "X-Content-Type-Options": {
            "$ref": "#/components/headers/NoSniff"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthError"
            }
          }
        }
      },
      "AuthThrottled": {
        "description": "Auth admission budget exhausted.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "private, no-store"
            }
          },
          "Referrer-Policy": {
            "schema": {
              "type": "string",
              "const": "no-referrer"
            }
          },
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/CorrelationId"
          },
          "X-Content-Type-Options": {
            "$ref": "#/components/headers/NoSniff"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthError"
            }
          }
        }
      },
      "UnexpectedFailure": {
        "description": "Safe unexpected failure; no internal details.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "private, no-store"
            }
          },
          "Referrer-Policy": {
            "schema": {
              "type": "string",
              "const": "no-referrer"
            }
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/CorrelationId"
          },
          "X-Content-Type-Options": {
            "$ref": "#/components/headers/NoSniff"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthError"
            }
          }
        }
      }
    }
  }
}